+++ lib/packages/core/src/api/composeApi.ts
@@ -29,6 +29,7 @@ import {getPeer} from './factories/network/getPeer';
 import {getPeers} from './factories/network/getPeers';
 import {sendMoney} from './factories/transaction/sendMoney';
 import {getAliases} from './factories/account/getAliases';
+import {getTime} from './factories/network/getTime';
 import {getAccount} from './factories/account/getAccount';
 
 
@@ -68,6 +69,7 @@ export function composeApi(settings: ApiSettings): Api {
             suggestFee,
             getPeers,
             getPeer,
+            getTime,
         })
         .withTransactionApi({
             broadcastTransaction,