+++ lib/packages/core/src/index.ts
@@ -1,6 +1,5 @@
 /** @module core */
 export * from './typings/api';
-export * from './typings/api/apiError';
 export * from './typings/account';
 export * from './typings/alias';
 export * from './typings/aliasList';
@@ -11,7 +10,7 @@ export * from './typings/blockchainStatus';
 export * from './typings/blockId';
 export * from './typings/burstNode';
 export * from './typings/peer';
-export * from './typings/peersResponse';
+export * from './typings/peerAddressList';
 export * from './typings/serverStatus';
 export * from './typings/suggestedFees';
 export * from './typings/transaction';
@@ -19,7 +18,7 @@ export * from './typings/transactionId';
 export * from './typings/transactionList';
 export * from './typings/transactionResponse';
 export * from './typings/unconfirmedTransactionList';
-export * from './typings/currency';
+export * from './typings/burstTime';
 export * from './api';