+++ lib/packages/core/src/typings/burstTime.ts
@@ -1,0 +1,12 @@
+/** @module core */
+
+/**
+ * Original work Copyright (c) 2019 Burst Apps Team
+ */
+
+export interface BurstTime {
+    /**
+     * The burst time (in seconds since the genesis block)
+     */
+    readonly time: number;
+}