+++ web/angular-wallet/src/app/layout/components/navbar/vertical/style-1/style-1.component.ts
@@ -42,8 +42,8 @@ export class NavbarVerticalStyle1Component implements OnInit, OnDestroy
         private _fuseConfigService: FuseConfigService,
         private _fuseNavigationService: FuseNavigationService,
         private _fuseSidebarService: FuseSidebarService,
-        private _storeService: StoreService,
         private _accountService: AccountService,
+        private _storeService: StoreService,
         private _router: Router
     )
     {