+++ lib/packages/core/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.1.0-rc.3
+- BRS exceptions thrown as `HttpError` now
+- Reduced lib size
+- added API functions
+    - Network: `getTime`
+    - Network: `getPeer`
+    - Network: `getPeers`
+    - Block: `getBlocks`
+    
+ 
 
 ## Unreleased
 ### Account